[export] Fix bazel protoc actions with `--experimental_sibling_repository_layout`
Protoc actions are failing when compiling protobuf in the `cc|objc_grpc_library` rules if `--experimental_sibling_repository_layout` is enabled. The assumptions made when generating certain paths are no longer true if sibling layout is used.

These fixes are based on observations of how sibling repositories are laid out.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/636309539](http://cl/636309539)

PiperOrigin-RevId: 636309539